src/requirements.txt
docx2txt
argparse==1.4.0
python-docx